UCS: move empty function stubs to the separate header file
