+++ src/tools/perf/lib/libperf.c
@@ -10,6 +10,7 @@
 #include <ucs/debug/log.h>
 #include <ucs/arch/bitops.h>
 #include <ucs/sys/module.h>
+
 #include <string.h>
 #include <malloc.h>
 #include <tools/perf/lib/libperf_int.h>
