+++ src/ucs/sys/compiler_def.h
@@ -135,7 +135,7 @@
  * @param _type   Structure type.
  * @param _field  Field of structure.
  * 
- * @return Size of _member in _type.
+ * @return Size of _field in _type.
  */
 #define ucs_field_sizeof(_type, _field) \
     sizeof(((_type*)0)->_field)
@@ -144,7 +144,7 @@
  * @param _type   Structure type.
  * @param _field  Field of structure.
  *
- * @return Type of _member in _type.
+ * @return Type of _field in _type.
  */
 #define ucs_field_type(_type, _field) \
     typeof(((_type*)0)->_field)
