+++ src/ucs/sys/sys.c
@@ -994,63 +994,3 @@ void ucs_sys_free(void *ptr, size_t length)
         }
     }
 }
-
-void ucs_empty_function()
-{
-}
-
-unsigned ucs_empty_function_return_zero()
-{
-    return 0;
-}
-
-int64_t ucs_empty_function_return_zero_int64()
-{
-    return 0;
-}
-
-ucs_status_t ucs_empty_function_return_success()
-{
-    return UCS_OK;
-}
-
-ucs_status_t ucs_empty_function_return_unsupported()
-{
-    return UCS_ERR_UNSUPPORTED;
-}
-
-ucs_status_t ucs_empty_function_return_inprogress()
-{
-    return UCS_INPROGRESS;
-}
-
-ucs_status_t ucs_empty_function_return_no_resource()
-{
-    return UCS_ERR_NO_RESOURCE;
-}
-
-ucs_status_ptr_t ucs_empty_function_return_ptr_no_resource()
-{
-    return UCS_STATUS_PTR(UCS_ERR_NO_RESOURCE);
-}
-
-ucs_status_t ucs_empty_function_return_ep_timeout()
-{
-    return UCS_ERR_ENDPOINT_TIMEOUT;
-}
-
-ssize_t ucs_empty_function_return_bc_ep_timeout()
-{
-    return UCS_ERR_ENDPOINT_TIMEOUT;
-}
-
-ucs_status_t ucs_empty_function_return_busy()
-{
-    return UCS_ERR_BUSY;
-}
-
-int ucs_empty_function_do_assert()
-{
-    ucs_assert_always(0);
-    return 0;
-}
